The start of a pipelined architecture
Instructions are now actual entities in the design. The hope is that
in the future these will be fleshed out to enable an actual pipelined
simulator.

Instructions come with some niceties: We can now easily convert between
a u16 (the instruction in 'binary' form) and a readable version of the
instruction. This is done by implementing the From trait for u16 and
the Instruction enum.

The Simulator now decodes the instruction (which is a simple
Instruction::from call), and then executes this instruction.

I've also moved the check inside Simulator::trace to before we call
the function as it makes it less complicated (and hopefully a very minor
optimisation in the process).